vpcd: added more input checking (#328)
May fix unstability issues #326 #324

+++ virtualsmartcard/win32/BixVReader/memory.h
@@ -3,8 +3,8 @@
 #include "device.h"
 
 
-bool getBuffer(IWDFIoRequest* pRequest,void **buffer,int *bufferLen);
-void setString(CMyDevice *device,IWDFIoRequest* pRequest,char *result,int outSize);
-void setBuffer(CMyDevice *device,IWDFIoRequest* pRequest,void *result,int inSize);
+bool getBuffer(IWDFIoRequest* pRequest,void **buffer,SIZE_T *bufferLen);
+void setString(CMyDevice *device,IWDFIoRequest* pRequest,char *result,SIZE_T outSize);
+void setBuffer(CMyDevice *device,IWDFIoRequest* pRequest,void *result,SIZE_T inSize);
 void setInt(CMyDevice *device,IWDFIoRequest* pRequest,DWORD result);
 DWORD getInt(IWDFIoRequest* pRequest);
